requirements.txt
kafka-python
praw
flask
